Add TechInfo object. JSON editing for both.
I reorganized the code so that the data handling is all done through a
TechInfo object. I tried to be a little more friendly to the Dependency
Injection pattern and reduce the use of globals.

I made the JSON fiddling button creation more dynamic and added it to
both User Groups and Events.

I moved ToggleClass to its own file.